+++ backend/canisters/community/impl/src/lifecycle/post_upgrade.rs
@@ -1,7 +1,7 @@
 use crate::jobs::import_groups::finalize_group_import;
 use crate::lifecycle::{init_env, init_state};
 use crate::memory::get_upgrades_memory;
-use crate::{read_state, Data};
+use crate::{mutate_state, read_state, Data};
 use canister_logger::LogEntry;
 use canister_tracing_macros::trace;
 use community_canister::post_upgrade::Args;
@@ -9,6 +9,7 @@ use ic_cdk::post_upgrade;
 use instruction_counts_log::InstructionCountFunctionId;
 use stable_memory::get_reader;
 use tracing::info;
+use types::Timestamped;
 
 #[post_upgrade]
 #[trace]
@@ -36,5 +37,15 @@ fn post_upgrade(args: Args) {
         state
             .data
             .record_instructions_count(InstructionCountFunctionId::PostUpgrade, now)
+    });
+
+    // TODO: Delete this one-time code
+    mutate_state(|state| {
+        let now = state.env.now();
+        for channel in state.data.channels.iter_mut() {
+            if channel.chat.is_public.value && channel.chat.gate.value.is_none() {
+                channel.chat.messages_visible_to_non_members = Timestamped::new(true, now);
+            }
+        }
     });
 }