+++ backend/canisters/community/impl/src/updates/create_channel.rs
@@ -75,6 +75,8 @@ fn create_channel_impl(
         return CommunityFrozen;
     }
 
+    let messages_visible_to_non_members = args.is_public && args.messages_visible_to_non_members.unwrap_or(args.gate.is_none());
+
     let caller = state.env.caller();
     if let Some(member) = state.data.members.get_mut(caller) {
         if member.suspended.value {
@@ -129,6 +131,7 @@ fn create_channel_impl(
                 subtype,
                 args.avatar,
                 args.history_visible_to_new_joiners,
+                messages_visible_to_non_members,
                 permissions,
                 args.gate.clone(),
                 args.events_ttl,