+++ backend/canisters/group/impl/src/updates/update_group_v2.rs
@@ -150,6 +150,7 @@ fn commit(my_user_id: UserId, args: Args, state: &mut RuntimeState) -> SuccessRe
         args.permissions_v2,
         args.gate,
         args.public,
+        args.messages_visible_to_non_members,
         args.events_ttl,
         state.env.now(),
     );