+++ backend/canisters/group_index/api/src/updates/c2c_create_group.rs
@@ -11,6 +11,8 @@ pub struct Args {
     pub subtype: Option<GroupSubtype>,
     pub avatar: Option<Document>,
     pub history_visible_to_new_joiners: bool,
+    #[serde(default)]
+    pub messages_visible_to_non_members: Option<bool>,
     pub permissions_v2: Option<GroupPermissions>,
     pub events_ttl: Option<Milliseconds>,
     pub gate: Option<AccessGate>,