+++ backend/integration_tests/src/communities/disappearing_message_tests.rs
@@ -33,6 +33,7 @@ fn disappearing_messages_in_channel() {
             events_ttl: OptionUpdate::SetToSome(1000),
             gate: OptionUpdate::NoChange,
             public: None,
+            messages_visible_to_non_members: None,
         },
     );
 
@@ -78,6 +79,7 @@ fn disappearing_messages_in_channel() {
             events_ttl: OptionUpdate::SetToNone,
             gate: OptionUpdate::NoChange,
             public: None,
+            messages_visible_to_non_members: None,
         },
     );