+++ backend/integration_tests/src/communities/update_channel_tests.rs
@@ -72,6 +72,7 @@ fn members_added_if_channel_made_public_or_gate_removed(make_public: bool) {
             events_ttl: OptionUpdate::NoChange,
             gate: if !make_public { OptionUpdate::SetToNone } else { OptionUpdate::NoChange },
             public: make_public.then_some(true),
+            messages_visible_to_non_members: None,
         },
     );